src/requirements.txt
# Core dependencies
functions-framework==3.*
flask==3.0.0
requests==2.31.0
tenacity==8.2.3

google-cloud-core
# Google Cloud dependencies
google-cloud-bigquery==3.12.0
google-cloud-storage==2.10.0
google-api-python-client==2.104.0

# Development utilities
python-dotenv==1.0.0